Parallel CV (uses CV_GLASSOc)

Parallel implementation of cross validation.

CVP_GLASSO(X = NULL, lam = 10^seq(-2, 2, 0.2), diagonal = FALSE,
  crit.out = c("avg", "max"), crit.in = c("loss", "avg", "max"),
  tol.out = 1e-04, tol.in = 1e-04, maxit.out = 10000, maxit.in = 10000,
  adjmaxit.out = NULL, K = 5, crit.cv = c("loglik", "AIC", "BIC"),
  start = c("warm", "cold"), cores = 1, trace = c("progress", "print",
  "none"))

Arguments

X
nxp data matrix. Each row corresponds to a single observation and each column contains n observations of a single feature/variable.

lam
positive tuning parameters for elastic net penalty. If a vector of parameters is provided, they should be in increasing order. Defaults to grid of values 10^seq(-2, 2, 0.2).

diagonal
option to penalize the diagonal elements of the estimated precision matrix (\(\Omega\)). Defaults to FALSE.

crit.out
criterion for convergence in outer (blockwise) loop. Criterion avg will loop until the average absolute parameter change is less than tol.out times tolerance multiple. Criterion max will loop until the maximum change in the estimated Sigma after an iteration over the parameter set is less than tol.out. Defaults to avg.

crit.in
criterion for convergence in inner (lasso) loop. Criterion for convergence. Criterion loss will loop until the relative change in the objective for each response after an iteration is less than tol.in. Criterion avg will loop until the average absolute change for each response is less than tol.in times tolerance multiple. Similary, criterion max will loop until the maximum absolute change is less than tol.in times tolerance multiple. Defaults to loss.

tol.out
convergence tolerance for outer (blockwise) loop. Defaults to 1e-4.

tol.in
convergence tolerance for inner (lasso) loop. Defaults to 1e-4.

maxit.out
maximum number of iterations for outer (blockwise) loop. Defaults to 1e4.

maxit.in
maximum number of iterations for inner (lasso) loop. Defaults to 1e4.

adjmaxit.out
adjusted maximum number of iterations. During cross validation this option allows the user to adjust the maximum number of iterations after the first lam tuning parameter has converged (for each alpha). This option is intended to be paired with warm starts and allows for 'one-step' estimators. Defaults to NULL.

K
specify the number of folds for cross validation.

crit.cv
cross validation criterion (loglik, AIC, or BIC). Defaults to loglik.

start
specify warm or cold start for cross validation. Default is warm.

cores
option to run CV in parallel. Defaults to cores = 1.

trace
option to display progress of CV. Choose one of progress to print a progress bar, print to print completed tuning parameters, or none.

Value

returns list of returns which includes:

lam

optimal tuning parameter.

min.error

minimum average cross validation error (cv.crit) for optimal parameters.

avg.error

average cross validation error (cv.crit) across all folds.

cv.error

cross validation errors (cv.crit).