CV penalized precision matrix estimation (c++)

Cross validation function for GLASSO.

CV_GLASSOc(X, S, lam, diagonal = FALSE, path = FALSE, crit_out = "avg",
  crit_in = "loss", tol_out = 1e-04, tol_in = 1e-04, maxit_out = 10000L,
  maxit_in = 10000L, adjmaxit_out = 10000L, K = 5L, crit_cv = "loglik",
  start = "warm", trace = "progress")

Arguments

X
option to provide a nxp matrix. Each row corresponds to a single observation and each column contains n observations of a single feature/variable.

S
option to provide a pxp sample covariance matrix (denominator n). If argument is NULL and X is provided instead then S will be computed automatically.

lam
positive tuning parameters for elastic net penalty. If a vector of parameters is provided, they should be in increasing order.

diagonal
option to penalize the diagonal elements of the estimated precision matrix (\(\Omega\)). Defaults to FALSE.

path
option to return the regularization path. This option should be used with extreme care if the dimension is large. If set to TRUE, cores will be set to 1 and errors and optimal tuning parameters will based on the full sample. Defaults to FALSE.

crit_out
criterion for convergence in outer (blockwise) loop. Criterion avg will loop until the average absolute parameter change is less than tol_out times tolerance multiple. Criterion max will loop until the maximum change in the estimated Sigma after an iteration over the parameter set is less than tol_out. Defaults to avg.

crit_in
criterion for convergence in inner (lasso) loop. Criterion for convergence. Criterion loss will loop until the relative change in the objective for each response after an iteration is less than tol_in. Criterion avg will loop until the average absolute change for each response is less than tol_in times tolerance multiple. Similary, criterion max will loop until the maximum absolute change is less than tol_in times tolerance multiple. Defaults to loss.

tol_out
convergence tolerance for outer (blockwise) loop. Defaults to 1e-4.

tol_in
convergence tolerance for inner (lasso) loop. Defaults to 1e-4.

maxit_out
maximum number of iterations for outer (blockwise) loop. Defaults to 1e4.

maxit_in
maximum number of iterations for inner (lasso) loop. Defaults to 1e4.

adjmaxit_out
adjusted maximum number of iterations. During cross validation this option allows the user to adjust the maximum number of iterations after the first lam tuning parameter has converged (for each alpha). This option is intended to be paired with warm starts and allows for "one-step" estimators. Defaults to 1e4.

K
specify the number of folds for cross validation.

crit_cv
cross validation criterion (loglik, AIC, or BIC). Defaults to loglik.

start
specify warm or cold start for cross validation. Default is warm.

trace
option to display progress of CV. Choose one of progress to print a progress bar, print to print completed tuning parameters, or none.

Value

list of returns includes:

lam

optimal tuning parameter.

path

array containing the solution path. Solutions will be ordered by ascending lambda values.

min.error

minimum average cross validation error (cv.crit) for optimal parameters.

avg.error

average cross validation error (cv.crit) across all folds.

cv.error

cross validation errors (cv.crit).